# webserver/swagger.yml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: ML Aide
  license:
    name: unspecified
servers:
  - url: http://api.mlaide.io/v1
security:
  - oauth2: []
paths:
...
  /users/current:
    get:
      summary: Get current user
      operationId: getCurrentUser
      tags:
        - users
      responses:
        '200':
          description: Current user object
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      summary: Update current user
      operationId: updateCurrentUser
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        '204':
          description: User updated
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/current/api-keys:
    get:
      summary: Get API keys of current user
      operationId: getApiKeys
      tags:
        - users
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeys"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a new API key
      operationId: createApiKey
      tags:
        - users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApiKey"
      responses:
        '200':
          description: API key created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKey"
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /users/current/api-keys/{apiKeyId}:
    delete:
      summary: Delete API key
      operationId: deleteApiKey
      tags:
        - users
      parameters:
        - name: apiKeyId
          in: path
          required: true
          description: The id of the api key
          schema:
            type: string
      responses:
        '204':
          description: Delete was successful
        '500':
          description: unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    ApiKey:
      type: object
      required:
        - createdAt
        - id
      properties:
        apiKey:
          type: string
          description: This property is only set in the response of the createApiKey operation
          readOnly: true
        createdAt:
          type: string
          format: date-time
          readOnly: true
        description:
          type: string
        expiresAt:
          type: string
          format: date-time
        id:
          type: string
          readOnly: true
    ApiKeys:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ApiKey"
...
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: integer
          format: int32
          example: 404
        message:
          type: string
          example: The requested resource does not exist
...
    User:
      type: object
      readOnly: true
      required:
        - email
        - id
        - nickName
      properties:
        userId:
          type: string
          readOnly: true
          example: 5def26a1-a932-404a-998c-aabf84994dd0
        email:
          type: string
          readOnly: true
          example: adam-test@mlaide.com
        firstName:
          type: string
          readOnly: false
          example: Adam
        lastName:
          type: string
          readOnly: false
          example: Test
        nickName:
          type: string
          readOnly: false
          example: Alice
...
  securitySchemes:
    oauth2:
      type: oauth2
      description: This API uses OAuth 2 with the Authorization Code Flow with Proof Key for Code Exchange (PKCE).
      flows:
        authorizationCode:
          authorizationUrl: https://mlaide-dev.eu.auth0.com/authorize
          tokenUrl: https://mlaide-dev.eu.auth0.com/oauth/token
          scopes:
            projects:create: Create new projects
            projects:edit: Edit existing projects